+++ CHANGELOG.md
@@ -18,6 +18,10 @@ Documentation for TransferBench is available at
 - Added new bmasweep preset that compares DMA to batched DMA execution for parallel transfers to other GPUs
 - Added new wallclock preset that compares wallclock counters across XCCs within a GPU
 - Added new smoketest preset that runs a variety of DMA/GFX tests for simple correctness tests
+- Added new help preset to show config file examples
+- Added new presets preset to show available presets and their descriptions
+- Added new envvars preset to show environment variables that can change TransferBench behavior
+- Adding information on how to run multi-rank with TransferBench, when run with no args
 
 ### Modified
 - DMA-BUF support enablement in CMake changed to ENABLE_DMA_BUF to be more similar to other compile-time options
@@ -30,6 +34,7 @@ Documentation for TransferBench is available at
 - Dynamically size CQs for NIC transfers in high QPs case
 - Switch to using hipMemcpyDeviceToDeviceNoCU instead of hipMemcpyDefault for DMA Executor if available (requires HIP >= 6.0)
 - Allow for multiple destination memory locations for DMA/Batched-DMA Transfers
+- Removed env vars printing and preset print when running TransferBench with no args
 
 ## v1.66.02
 ### Added